requirements.txt
openai==0.28
Telethon
python-dotenv
pymysql
cryptography